% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB_DA_test.R
\name{BB_DA_test}
\alias{BB_DA_test}
\title{Differential abundance analysis using a Beta Binomial (BB) based model}
\usage{
BB_DA_test(Count_vector, Total_cell, Condition_vector, type_test = "LRT")
}
\arguments{
\item{Count_vector}{a numerical vector containing the list of count values for a given cell of interest.}

\item{Total_cell}{Total number of cells in each FoV}

\item{Condition_vector}{a character vector of factor of the same size as Count_vector that describes to which group each sample belongs.}

\item{type_test}{the type of statistical test to perform in order to compute the statistical significance. Has to be either LRT (Likelihood Ratio Test) or Wald (Wald's test)}
}
\value{
P-value for differential abundance between the groups of sample. If LRT has been selected only one p-value will be provided. If Wald test was selected one p-value per condition - 1 will be provided
}
\description{
Performs a differential abundance analysis between groups of FoVs. Unlike the NB based test, the FoVs do not need to have the same size.
}
\examples{
Count_vector = c(rnbinom(n=100,mu=2,size=1),rnbinom(n=50,mu=6,size=1))
Condition_vector = c(rep("Control",100),rep("Treatment",50))
Total_cell = rep(200,150)
BB_DA_test(Count_vector,Total_cell,Condition_vector)
}
