% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_QC.R
\name{Clustering_QC}
\alias{Clustering_QC}
\title{Comparison and QC of clustering result}
\usage{
Clustering_QC(sce, assay_type = "Raw_intensity")
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay_type}{name of the data slot on which to perform the clustering (Raw_intensity, Arcsinh_transformed_intensity....). By default the regression-normalized data are used}
}
\value{
Returns a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with a new assay slot called "Count_normalised_intensity"
}
\description{
This function controls the quality of the clustering by comparing cluster
}
\examples{
sce = Count_normalization(sce,residual_normalisation = "Anscombe")
}
