% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma_transformation.R
\name{Gamma_transformation}
\alias{Gamma_transformation}
\title{Normalisation and scaling of marker expression based on Wilson-Hilferty transform}
\usage{
Gamma_transformation(sce, clip_quantile = NULL)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{clip_quantile}{the percentile used to clip intensity values.}
}
\value{
Returns a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with a new assay slot called "Count_normalised_intensity"
}
\description{
This function normalize and transform the intensity using the Wilson-Hilferty transform (cube root transform) and then scale the intensity
}
\examples{
sce = Gamma_transformation(sce,clip_quantile=0.001)
}
