% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_alpha_estimation.R
\name{Global_alpha_estimation}
\alias{Global_alpha_estimation}
\title{Estimation of alpha using many small FoVs}
\usage{
Global_alpha_estimation(sce, List_image = NULL, Threshold_detection = 20)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{List_image}{a vector with the number of the images to use for the estimate. If not provided, all images will be used.}

\item{Threshold_detection}{minimal number of cells required to consider a cell type/cluster as detected. By default set to 20.}

\item{Value_split}{vector containing the fraction of the FoV that are taken, i.e numbers bigger than 1. If not provided the following vector will be used : 1, 1.1, 1.2, 1.5, 1.8, 2, 2.5 and 3.}
}
\value{
Returns a data.frame with the estimate of alpha for each FoV as well as the quality of each fitting (R-squared).
}
\description{
Function that estimates the alpha parameter of a given tissue using many small FoVs instead of a unique large FoVs.
}
\examples{
#Simple case where only the number of sample regions changes :
Alpha_estimate = Global_alpha_estimation(sce)
}
