% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_clustering.R
\name{KNN_clustering}
\alias{KNN_clustering}
\title{Perform cell clustering on a computed KNN graph}
\usage{
KNN_clustering(
  sce,
  K = 30,
  clustering_method = "Louvain",
  assay_type = "Count_normalised_intensity",
  metric = "angular"
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{K}{number of neighbors for the KNN graph computation}

\item{clustering_method}{method used for graph clustering. Has to be chose among "Louvain","Greedy" and "Infomap"}

\item{assay_type}{name of the data slot on which to perform the clustering (Raw_intensity, Arcsinh_transformed_intensity....). By default the regression-normalized data are used}
}
\value{
Returns an updated \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with an updated colLabels slot
}
\description{
Perform cell clustering on a computed KNN graph
}
\examples{
sce = KNN_clustering(sce,K = 30,clustering_method = "Louvain")
#'
}
