% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MD_index_tensor.R
\name{Compute_bivariate_MD_index}
\alias{Compute_bivariate_MD_index}
\title{Compute the mean direction index}
\usage{
Compute_bivariate_MD_index(
  X_pattern,
  Y_pattern,
  k_parameter = 4,
  normalize_MD = TRUE
)
}
\arguments{
\item{k_parameter}{number of NN to use to compute the mean direction index}

\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}
}
\value{
Returns three-order tensor with the log2 CE index of each pair of cell type in each image
}
\description{
Compute the mean direction (MD) index for each pair of cell type in each image and store them in a three-order tensor
}
\examples{
MD_tensor = MD_index_tensor(sce)
}
