% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVS_thresholding.R
\name{MVS_thresholding}
\alias{MVS_thresholding}
\title{Minimum Variance Stratification thresholding}
\usage{
MVS_thresholding(x, L = 6, distribution = "beta")
}
\arguments{
\item{x}{numeric vector to be stratified}

\item{L}{number of strata. Should be between 4 and 6}

\item{distribution}{Probability distribution model used for the thresholding. So far only the beta distribution is available.}
}
\value{
Returns a vector containing the L-1 thresholding values and produces a plot showing the threshold values.
}
\description{
Function that computes the optimal thresholding of a numeric vector into L strata using the Minimum Variance Stratification (MVS) thresholding. For a complete description of the MVS thresholding strategy please have a look the paper "Minimum Variance Stratification" by Tore Dalenius and Joseph L. Hodge in 1959
}
\examples{
x = rbeta(n = 1000,shape1 = 2,shape2 = 5)
MVS_thresholding(x,L=6)
}
