% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch_detection.R
\name{Patch_detection}
\alias{Patch_detection}
\title{Cellular patch detection}
\usage{
Patch_detection(
  sce,
  graph_type = "KNN",
  graph_parameter = NULL,
  neighbor_degree = 2,
  clustering_method = "saturation",
  distance_composition = "Bray-Curtis",
  min_points = 15,
  image_list = NULL
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{graph_type}{the type of graph to use to compute cellular neighborhood. Can be "KNN" for a k-nearest neighbor , "Radius" for epsilon-radius graph or  "Gabriel for a Gabriel graph.}

\item{graph_parameter}{the parameter value for the graph construction. Corresponds to the number of neighbor for the KNN and the radius value for the epsilon-radius graph.}

\item{neighbor_degree}{the degree of the neigbor used to compute the similarity of cellular environement between two cells. Typically 1 or 2.}

\item{clustering_method}{method used to clusters the cells once similarity/distances have been computed}

\item{distance_composition}{type of distance to use to compare neighbor composition. Can be "cosine" for cosine similarity, "JS" for Jensen-Shanon divergence}

\item{min_points}{minimal number of points to consider a community of points a patch}

\item{image_list}{images/FOV/location on which to perform the analysis. If not provided the analysis is done on all images.}
}
\value{

}
\description{
This function identifies patches of cells, i.e physically neighbor cells that share a similar cellular environement
}
\examples{
sce = Patch_detection(sce,graph_type="KNN",graph_parameter=30,neighbor_degree=2)
}
