% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_sampling_analysis.R
\name{Perform_sampling_analysis}
\alias{Perform_sampling_analysis}
\title{Sampling analysis}
\usage{
Perform_sampling_analysis(
  sce,
  Selected_image = 1,
  N_times = 50,
  N_sampling_region_vector = 1:20,
  width_FOV_vector = 400,
  height_FOV_vector = 400,
  Threshold_detection_cluster = 50
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{Selected_image}{of which image/FOV should the sampling analysis be performed ?}

\item{N_times}{number of times each type of sampling is performed (typically between 20 and 100)}

\item{N_sampling_region_vector}{vector (or integer) describing the different values taken by the number of sampled regions}

\item{width_FOV_vector}{vector (or real number) describing the width of the FOV}

\item{height_FOV_vector}{vector (or real number) describing the height of the FOV}

\item{Threshold_detection_cluster}{real number corresponding to the minimal number of a given cell type to be considered as detected. Typically around 50 or 100.}
}
\value{
Returns a data.frame containing various summary statistics such as the mean number of detected clusters, the KL divergence with the real cell type composition etc...
}
\description{
Perform a sampling analysis where one or various sampling parameters are varying
}
\examples{
#Simple case where only the number of sample regions changes :
Perform_sampling_analysis(sce,Selected_image=1,N_times=50,N_sampling_region_vector=1:20,width_FOV_vector=400,height_FOV_vector=400,Threshold_detection=50)
}
