% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_stratified_sampling.R,
%   R/Perform_stratified_sampling_corrected.R
\name{Perform_stratified_sampling_simulation}
\alias{Perform_stratified_sampling_simulation}
\alias{Perform_stratified_sampling_simulation_2}
\title{Perform stratified sampling simulation for a list of markers}
\usage{
Perform_stratified_sampling_simulation(
  sce,
  tiff_file,
  panel_file,
  N_simulations = 30,
  N_FoV = 10,
  FoV_size = 100,
  type_stratification = "Neyman",
  L = 6,
  sigma = 100,
  specific_channels = NULL,
  perform_rotation = TRUE,
  show_plot = TRUE,
  Parallel_computing = TRUE
)

Perform_stratified_sampling_simulation_2(
  sce,
  tiff_file,
  panel_file,
  N_simulations = 30,
  N_FoV = 10,
  FoV_size = 100,
  type_stratification = "Neyman",
  type_thresholding = "Cumsqrt",
  L = 6,
  sigma = 100,
  specific_channels = NULL,
  perform_rotation = TRUE,
  show_plot = TRUE,
  Parallel_computing = TRUE
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{tiff_file}{path toward the raw tiff file}

\item{panel_file}{path toward a csv file associating each channel of the tiff file to gene. The file should have column with name "Target"}

\item{N_FoV}{Number of fields of views}

\item{type_stratification}{how should the allocation of FoVs across strata be computed ? Can be "Neyman" or "Proportional" for Neyman/optimal and proportional allocation respectively}

\item{L}{number of strata. Should ideally be between 4 and 6}

\item{sigma}{smoothing parameter}

\item{perform_rotation}{performing rotation between the sce object and tiff image coordinate}

\item{show_plot}{should the intermediate plots be shown, i.e results of tissue stratification. Boolean.}

\item{Parallel_computing}{should the different sampling be performed in parallel ? Boolean.}

\item{type_thresholding}{how should stratum be computed based on the intensity of the covariate ? Can be "Cumsqrt" or "Geometric"}

\item{Fov_size}{Size (i.e width/height) of the square FoVs}

\item{specific_channel}{which channels should be tested for stratified sampling ? If not provided all channels will be used}
}
\value{
A list of table, each table corresponding to a different stratified sampling

A list of table, each table corresponding to a different stratified sampling
}
\description{
Allows to perform stratified sampling simulation for a list of markers.

Allows to perform stratified sampling simulation for a list of markers.
}
