% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_cluster_gene_expression.R
\name{Plot_cluster_gene_expression}
\alias{Plot_cluster_gene_expression}
\title{Visualization of a specific gene expression across clusters}
\usage{
Plot_cluster_gene_expression(
  sce,
  assay_type = "Count_normalised_intensity",
  Gene = NULL
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay_type}{name of the data slot on which to perform the clustering (Raw_intensity, Arcsinh_transformed_intensity....). By default the regression-normalized data are used}

\item{Gene}{Name of the gene to be plotted}
}
\value{
Return a plot
}
\description{
Plot a boxplot of gene expression across clusters
}
\examples{
Plot_cluster_gene_expression(sce,Gene = "CD45")


}
