% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_cluster_spatial.R
\name{Plot_cluster_spatial}
\alias{Plot_cluster_spatial}
\title{Visualization of cluster spatial distribution}
\usage{
Plot_cluster_spatial(
  sce,
  Image_number = 1,
  Cex_parameter = 10,
  Specific_cluster = NULL,
  Provided_cluster = NULL
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}. Clustering of the cell should have been previously performed.}

\item{Image_number}{Number of name of the image/ROI to be plotted}

\item{Cex_parameter}{Scaling factor for the size of the cells}

\item{Specific_cluster}{Colors points based based on their belonging to a specific cluster}

\item{Provided_group}{User-provided clustering}
}
\value{
Return a plot
}
\description{
Plot the spatial distribution of the identified clusters or of any other provided grouping of the cells
}
\examples{
Plot_cluster_spatial(sce,Image_number = 1)


}
