% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region_guided_sampling.R
\name{Region_guided_sampling}
\alias{Region_guided_sampling}
\title{Random sampling within a specific part of the tissue.}
\usage{
Region_guided_sampling(
  Segmentation_image,
  Selected_regions = 1,
  N_FoV = 1,
  FoV_size = 100,
  show_plot = FALSE
)
}
\arguments{
\item{Segmentation_image}{a cimg object}

\item{Selected_regions}{number of the stratum to be sampled.}

\item{N_FoV}{number of Field of Views (FoVs) to be sampled within the region of interest}

\item{FoV_size}{size of Field of Views (FoVs) to be sampled within the region of interest}

\item{show_plot}{boolean object. Should the result of the sampling be shown ?}
}
\value{
Returns the vector of all FoV center.
}
\description{
Random sampling within a specific part of the tissue using a thresholded image, typically done by MVS.
}
\examples{
Region_guided_sampling(MVS_thresholded_image,Selected_region = 1,N_FoV = 10,FoV_size = 200,show_plot = T)
}
