% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_round.R
\name{smart_round}
\alias{smart_round}
\title{Smart rounding of non integer vector}
\usage{
smart_round(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A rounded vector which sum is the same as the initial vector
}
\description{
Perform rounding of a numeric vector which conservs the total sum
}
