% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strata_thresholding.R
\name{Strata_thresholding}
\alias{Strata_thresholding}
\title{Computing thresholding values for}
\usage{
Strata_thresholding(x, L = 6, type_thresholding = "Cumsqrt")
}
\arguments{
\item{x}{numeric vector to be stratified}

\item{L}{number of strata. Should be between 4 and 6}

\item{type_thresholding}{how should stratum be computed based on the intensity of the covariate ? Can be "Cumsqrt" or "Geometric"}
}
\value{
Returns a vector containing the L-1 thresholding values and produces a plot showing the threshold values.
}
\description{
Function that computes the optimal thresholding of a numeric vector into L strata using dufferent approaches including the 'Cumsqrt' or Minimum Variance Stratification method
(for a detailed review see the paper "Stratification of Skewed Populations: A Review")
}
\examples{
x = rbeta(n = 1000,shape1 = 2,shape2 = 5)
MVS_thresholding(x,L=6)
}
