% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified_sampling.R
\name{Stratified_sampling}
\alias{Stratified_sampling}
\title{Simulation of a Stratified Sampling}
\usage{
Stratified_sampling(
  Thresholded_image,
  sce,
  N_FoV_per_region,
  FoV_size = 100,
  N_sampling = 50,
  Weight_strata = NULL,
  List_strata = NULL,
  Parallel_computing = FALSE,
  name_stratifying_marker = NULL
)
}
\arguments{
\item{Thresholded_image}{a cimg object}

\item{sce}{SingleCellExperiment object processed by balagan}

\item{N_FoV_per_region}{numeric vector. Number of FoV to sample in each stratum.}

\item{FoV_size}{size of Field of Views (FoVs) to be sampled.}

\item{N_sampling}{number of simulated sampling to perform.}

\item{Weight_strata}{numeric vector that must sum to 1, corresponding to the contribution of each stratum to the general tissue composition.}

\item{List_strata}{(optional) character or numeric vector. Ordered list of the stratum names.}

\item{Parallel_computing}{should the different sampling be performed in parallel ? Boolean.}

\item{name_stratifying_marker}{Name of the gene used for the stratification. Not essential, only for visualisation purpose}
}
\value{
Returns a table with the estimated cell density for each cell type and each simulation
}
\description{
Perform several simulation of a stratified sampling using a stratified image
}
\examples{
Stratified_sampling(Thresholded_image = MVS_thresholded_image,sce = sce,N_FoV_per_region = c(1,2,3,1,2),FoV_size = 200,N_sampling = 50 ,Weight_strata =c(0.1,0.2,0.4,0.2,0.1),List_strata = 1:6)
}
