% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP_embedding.R
\name{UMAP_embedding}
\alias{UMAP_embedding}
\title{Computation of a UMAP embedding}
\usage{
UMAP_embedding(
  sce,
  assay_type = "Count_normalised_intensity",
  metric = "correlation",
  n_neighbors = 30,
  use
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay_type}{name of the data slot on which to perform the clustering (Raw_intensity, Arcsinh_transformed_intensity....). By default the regression-normalized data are used}
}
\value{
Return an updated \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with an updated reducedDim slot.
}
\description{
Computes a UMAP embedding
}
\examples{
sce = UMAP_embedding(sce,assay_type = "Count_normalised_intensity", metric="correlation",n_neighbors = 30)


}
