% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_complex_sampling.R
\name{Visualize_complex_sampling}
\alias{Visualize_complex_sampling}
\title{Visualization of complex sampling analysis}
\usage{
Visualize_complex_sampling(Sampling_data_frame, Parameter_table)
}
\arguments{
\item{Sampling_data_frame}{a data.frame object resulting from the Perform_sampling_analysis() function}

\item{Parameter_table}{a table that describe the parameter}
}
\value{
Returns a plot as well as tau and No estimate for each of the parameter value.
}
\description{
Function that plot and analyze the results from the Perform_sampling_analysis() and estimate the tau and No parameters when various parameters value have been used .
}
\examples{
#Simple case where only the number of sample regions changes :
Visualize_sampling(Simple_sampling_analysis)
}
