% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_simple_sampling_KL.R
\name{Visualize_complex_sampling_KL}
\alias{Visualize_complex_sampling_KL}
\title{Visualization of complex sampling analysis focusing on cell composition}
\usage{
Visualize_complex_sampling_KL(Sampling_data_frame, Parameter_table)
}
\arguments{
\item{Sampling_data_frame}{a data.frame object resulting from the Perform_sampling_analysis() function}

\item{Parameter_table}{a table that describe the parameter}
}
\value{
Returns a plot as well as tau and No estimate for each of the parameter value.
}
\description{
Function that plot and analyze the results from the Perform_sampling_analysis() and estimate the and estimate the theta, KL_0 and KL_b parameters (Kullback-Leibler divergence analysis) when various parameters value have been used .
}
\examples{
#Simple case where only the number of sample regions changes :
Visualize_complex_sampling_KL(Simple_sampling_analysis)
}
