% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_CP.R
\name{Visualize_CP}
\alias{Visualize_CP}
\title{Visualize the result of a Canonical Polyadic (CP) decomposition}
\usage{
Visualize_CP(CP_decomposition, k = 1)
}
\arguments{
\item{CP_object}{the output of the cp (rTensor package) function}

\item{Plot_type}{the type of plot to produce. Has to be chosen among "all", "loading matrix" or}
}
\value{

}
\description{
Basic function allowing to visualize the result of a Canonical Polyadic (CP) decomposition
}
\examples{
Visualize_CP(CP_object)
}
