% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_simple_sampling.R
\name{Visualize_simple_sampling}
\alias{Visualize_simple_sampling}
\title{Visualization of simple sampling analysis}
\usage{
Visualize_simple_sampling(Sampling_data_frame)
}
\arguments{
\item{Sampling_data_frame}{a data.frame object resulting from the Perform_sampling_analysis() function}
}
\value{
Returns a plot as well as tau and No estimate for each of the parameter value.
}
\description{
Function that plot and analyze the results from the Perform_sampling_analysis() and estimate the tau and No parameters.
}
\examples{
#Simple case where only the number of sample regions changes :
Visualize_sampling(Simple_sampling_analysis)
}
