% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_stratified_sampling.R
\name{smart.round}
\alias{smart.round}
\title{Perform stratified sampling simulation for a list of markers}
\usage{
smart.round(x)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{tiff_file}{path toward the raw tiff file}

\item{panel_file}{path toward a csv file associating each channel of the tiff file to gene. The file should have column with name "Target"}

\item{L}{number of strata. Should ideally be between 4 and 6}

\item{N_FoV}{Number of fields of views}

\item{Fov_size}{Size (i.e width/height) of the square FoVs}

\item{sigma}{smoothing parameter}

\item{specific_channel}{which channels should be tested for stratified sampling ? If not provided all channels will be used}

\item{perform_rotation}{performing rotation between the sce object and tiff image coordinate}

\item{show_plot}{should the intermediate plots be shown, i.e results of tissue stratification. Boolean.}
}
\value{
A list of table, each table corresponding to a different stratified sampling
}
\description{
Allows to perform stratified sampling simulation for a list of markers.
}
